Prevent infinite loop in award fast-forward simulation
If judge result of a run is 'pending', a consistency between
runfeeder and team status gets broken (i.e. there is no more run
to feed for a team, but the team remains still pending),
which results in an infinite loop.

In such cases where run feeding fails, we just terminate the
simulation loop with an error message.